refactor: add pokemon list card
